ASoC: msm: qdsp6v2: Add a new waitqueue state for mem map/unmap
All of the commands being sent to the DSP are using cmd_state flag
which can be overwritten when several commands are sent at the same
time. This results in Memory map command timing out. The change will
separate the memory flag from flags for other commands so that they
do not overwrite each other.

Change-Id: If42ff7c34180eb0004c97ffc0eb656f0c30be0d8
CRs-fixed: 673781
Signed-off-by: Damir Didjusto <damird@codeaurora.org>
Signed-off-by: Gopikrishnaiah Anandan <agopik@codeaurora.org>